Add selectionType setting with checked and numbered options
This change adds a new selectionType setting that defaults to checked.
If selectionType is changed to numbered, icons are drawn with a number
instead of a checkmark. Numbering is based on the selection order and
begins at 1. On deselection, if the cell wasn't the most recently
selected, all higher numbered selections are decremented so that the
most recent selection always equals the selection count.